enable busy detector only for qw-run
