+++ quickwit/quickwit-cli/src/main.rs
@@ -29,7 +29,7 @@ use quickwit_cli::cli::{build_cli, CliCommand};
 #[cfg(feature = "jemalloc")]
 use quickwit_cli::jemalloc::start_jemalloc_metrics_loop;
 use quickwit_cli::{
-    QW_ENABLE_JAEGER_EXPORTER_ENV_KEY, QW_ENABLE_OPENTELEMETRY_OTLP_EXPORTER_ENV_KEY,
+    busy_detector, QW_ENABLE_JAEGER_EXPORTER_ENV_KEY, QW_ENABLE_OPENTELEMETRY_OTLP_EXPORTER_ENV_KEY,
 };
 use quickwit_common::RED_COLOR;
 use quickwit_serve::{quickwit_build_info, QuickwitBuildInfo};
@@ -166,75 +166,6 @@ async fn main_impl() -> anyhow::Result<()> {
     telemetry_handle.terminate_telemetry().await;
     global::shutdown_tracer_provider();
     std::process::exit(return_code)
-}
-
-mod busy_detector {
-    use std::sync::atomic::{AtomicU64, Ordering};
-    use std::time::Instant;
-
-    use once_cell::sync::Lazy;
-    use tracing::warn;
-
-    // we need that time reference to use an atomic and not a mutex for LAST_UNPARK
-    static TIME_REF: Lazy<Instant> = Lazy::new(Instant::now);
-
-    const ALLOWED_DELAY_MICROS: u64 = 5000;
-    const WARN_SUPPRESSION_MICROS: u64 = 30_000_000;
-
-    // LAST_UNPARK_TIMESTAMP and NEXT_WARN_TIMESTAMP are semantically micro-second
-    // precision timestamps, but we use atomics to allow accessing them without locks.
-    thread_local!(static LAST_UNPARK_TIMESTAMP: AtomicU64 = AtomicU64::new(0));
-    static NEXT_WARN_TIMESTAMP: AtomicU64 = AtomicU64::new(0);
-    static SUPPRESSED_WARN_COUNT: AtomicU64 = AtomicU64::new(0);
-
-    pub fn thread_unpark() {
-        LAST_UNPARK_TIMESTAMP.with(|time| {
-            let now = Instant::now()
-                .checked_duration_since(*TIME_REF)
-                .unwrap_or_default();
-            time.store(now.as_micros() as u64, Ordering::Relaxed);
-        })
-    }
-
-    pub fn thread_park() {
-        LAST_UNPARK_TIMESTAMP.with(|time| {
-            let now = Instant::now()
-                .checked_duration_since(*TIME_REF)
-                .unwrap_or_default();
-            let now = now.as_micros() as u64;
-            let delta = now - time.load(Ordering::Relaxed);
-            if delta > ALLOWED_DELAY_MICROS {
-                emit_warn(delta, now);
-            }
-        })
-    }
-
-    fn emit_warn(delta: u64, now: u64) {
-        if NEXT_WARN_TIMESTAMP
-            .fetch_update(Ordering::Relaxed, Ordering::Relaxed, |next_warn| {
-                if next_warn < now {
-                    Some(now + WARN_SUPPRESSION_MICROS)
-                } else {
-                    None
-                }
-            })
-            .is_err()
-        {
-            // a warn was emited recently, don't emit log for this one
-            SUPPRESSED_WARN_COUNT.fetch_add(1, Ordering::Relaxed);
-            return;
-        }
-
-        let suppressed = SUPPRESSED_WARN_COUNT.swap(0, Ordering::Relaxed);
-        if suppressed == 0 {
-            warn!("Thread wasn't parked for {delta}µs, is the runtime too busy?");
-        } else {
-            warn!(
-                "Thread wasn't parked for {delta}µs, is the runtime too busy? ({suppressed} \
-                 similar messages suppressed)"
-            );
-        }
-    }
 }
 
 /// Return the about text with telemetry info.
